Implements safe args
