+++ app/build.gradle
@@ -1,6 +1,7 @@
 plugins {
     id 'com.android.application'
     id 'org.jetbrains.kotlin.android'
+    id 'androidx.navigation.safeargs.kotlin'
 }
 
 android {
@@ -76,4 +77,7 @@ dependencies {
     androidTestImplementation 'androidx.test.espresso:espresso-contrib:3.5.1'
     testImplementation 'org.mockito:mockito-core:3.12.4'
     implementation 'org.threeten:threetenbp:1.6.3'
+
+
+
 }
