+++ app/src/main/res/navigation/nav_graph.xml
@@ -21,7 +21,10 @@
             app:destination="@id/searchFragment" />
         <action
             android:id="@+id/action_homeFragment_to_tomorrowFragment"
-            app:destination="@id/tomorrowFragment" />
+            app:destination="@id/tomorrowFragment"
+            app:popUpTo="@id/homeFragment"
+            app:popUpToSaveState="true"
+            />
 
     </fragment>
     <fragment
@@ -54,5 +57,9 @@
         <action
             android:id="@+id/action_tomorrowFragment_to_homeFragment"
             app:destination="@id/homeFragment" />
+        <argument
+            android:name="date"
+            app:argType="string"
+            android:defaultValue="tomorrow" />
     </fragment>
 </navigation>
